+++ src/app/_components/ShareModal.module.css
@@ -1,0 +1,31 @@
+.background {
+  background-color: rgba(0, 0, 0, 0.5);
+  top: 0;
+  bottom: 0;
+  right: 0;
+  left: 0;
+  position: absolute;
+  z-index: 101;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.container {
+  background-color: #2e2e2e;
+  color: white;
+  border-radius: 10px;
+  padding: 20px;
+  width: 300px;
+}
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border: 0px solid #505050;
+  padding-bottom: 10px;
+  border-bottom-width: 1px;
+}
+.title {
+  font-size: 16px;
+  font-weight: 500;
+}
