+++ src/app/_components/ShareModal.tsx
@@ -1,0 +1,1 @@
+const ShareModal = () => {};
